Updated to the new packages. Starting on openresty server install.

+++ packages/composer.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
-curl -sS https://getcomposer.org/installer | php
-mv composer.phar /usr/local/bin/composer
-print_success "Composer installed to /usr/local/bin/composer"
+if [[ ! -e /usr/local/bin/composer ]]; then
+	curl -sS https://getcomposer.org/installer | php
+	mv composer.phar /usr/local/bin/composer
+	print_success "Composer installed to /usr/local/bin/composer"
+fi
